backend/requirements.txt
# Django and Core Dependencies
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.7

# Real-time Communication
channels==4.0.0
channels-redis==4.2.0

# GraphQL
graphene-django==3.2.0

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# Machine Learning Integration
requests==2.31.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.4

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# API Documentation
drf-spectacular==0.27.0

# Utilities
Pillow==10.1.0
